+++ packages/fiori/src/NotificationList.ts
@@ -39,9 +39,26 @@ type NotificationItemCloseEventDetail = NotificationItemEventDetail;
  * The `ui5-notification-list` web component represents
  * a container for `ui5-li-notification-group` and `ui5-li-notification`.
  *
+ * ### Keyboard Handling
+ *
+ * #### Basic Navigation
+ * The `ui5-notification-list` provides advanced keyboard handling.
+ * When a list is focused the user can use the following keyboard
+ * shortcuts in order to perform a navigation:
+ *
+ * - [Up] or [Left] - Navigates up the items
+ * - [Down] or [Right] - Navigates down the items
+ * - [Home] - Navigates to first item
+ * - [End] - Navigates to the last item
+ *
+ * #### Fast Navigation
+ * This component provides a build in fast navigation group which can be used via [F6] / [Shift] + [F6] / [Ctrl] + [Alt/Option] / [Down] or [Ctrl] + [Alt/Option] + [Up].
+ * In order to use this functionality, you need to import the following module:
+ * `import "@ui5/webcomponents-base/dist/features/F6Navigation.js"`
+ *
  * ### ES6 Module Import
  *
- * `import "@ui5/webcomponents/dist/NotificationList.js";``
+ * `import "@ui5/webcomponents-fiori/dist/NotificationList.js";``
  * @constructor
  * @extends UI5Element
  * @since 2.0
@@ -114,11 +131,11 @@ class NotificationList extends UI5Element {
 
 	/**
 	 * Defines the text that is displayed when the component contains no items.
-	 * @default ""
+	 * @default undefined
 	 * @public
 	 */
 	@property()
-	noDataText!: string;
+	noDataText?: string;
 
 	get _accessibleName() {
 		return NotificationList.i18nFioriBundle.getText(NOTIFICATION_LIST_ACCESSIBLE_NAME);