+++ packages/main/src/ListItem.ts
@@ -5,7 +5,6 @@ import {
 } from "@ui5/webcomponents-base/dist/Keys.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import { getI18nBundle } from "@ui5/webcomponents-base/dist/i18nBundle.js";
-import { isDesktop } from "@ui5/webcomponents-base/dist/Device.js";
 import getActiveElement from "@ui5/webcomponents-base/dist/util/getActiveElement.js";
 import { getFirstFocusableElement } from "@ui5/webcomponents-base/dist/util/FocusableElements.js";
 import type { AccessibilityAttributes, PassiveEventListenerObject } from "@ui5/webcomponents-base/dist/types.js";
@@ -30,7 +29,7 @@ import {
 	LIST_ITEM_SELECTED,
 	LIST_ITEM_NOT_SELECTED,
 } from "./generated/i18n/i18n-defaults.js";
-import ListItemAccessibleRole from "./types/ListItemAccessibleRole.js";
+import type ListItemAccessibleRole from "./types/ListItemAccessibleRole.js";
 
 // Styles
 import styles from "./generated/themes/ListItem.css.js";
@@ -52,7 +51,7 @@ type SelectionRequestEventDetail = {
 }
 
 type AccInfo = {
-	role: string;
+	role?: string;
 	ariaExpanded?: boolean;
 	ariaLevel?: number;
 	ariaLabel: string;
@@ -109,8 +108,8 @@ abstract class ListItem extends ListItemBase {
 	 * @default "Active"
 	 * @public
 	*/
-	@property({ type: ListItemType, defaultValue: ListItemType.Active })
-	type!: `${ListItemType}`;
+	@property()
+	type: `${ListItemType}` = "Active";
 
 	/**
 	 * Defines the additional accessibility attributes that will be applied to the component.
@@ -127,7 +126,7 @@ abstract class ListItem extends ListItemBase {
 	 * @since 1.15.0
 	 */
 	@property({ type: Object })
-	accessibilityAttributes!: ListItemAccessibilityAttributes;
+	accessibilityAttributes: ListItemAccessibilityAttributes = {};
 
 	/**
 	 * The navigated state of the list item.
@@ -137,23 +136,23 @@ abstract class ListItem extends ListItemBase {
 	 * @since 1.10.0
 	 */
 	@property({ type: Boolean })
-	navigated!: boolean;
+	navigated = false;
 
 	/**
 	 * Defines the text of the tooltip that would be displayed for the list item.
-	 * @default ""
+	 * @default undefined
 	 * @public
 	 * @since 1.23.0
 	 */
-	@property({ type: String, defaultValue: "" })
-	tooltip!: string;
+	@property()
+	tooltip?: string;
 
 	/**
 	 * Indicates if the list item is active, e.g pressed down with the mouse or the keyboard keys.
 	 * @private
 	*/
 	@property({ type: Boolean })
-	active!: boolean;
+	active = false;
 
 	/**
 	 * Defines the highlight state of the list items.
@@ -162,8 +161,8 @@ abstract class ListItem extends ListItemBase {
 	 * @public
 	 * @since 1.24
 	 */
-	@property({ type: Highlight, defaultValue: Highlight.None })
-	highlight!: `${Highlight}`;
+	@property()
+	highlight: `${Highlight}` = "None";
 
 	/**
 	 * Used to define the role of the list item.
@@ -172,11 +171,11 @@ abstract class ListItem extends ListItemBase {
 	 * @since 1.3.0
 	 *
 	 */
-	@property({ type: ListItemAccessibleRole, defaultValue: ListItemAccessibleRole.ListItem })
-	accessibleRole!: `${ListItemAccessibleRole}`;
+	@property()
+	accessibleRole: `${ListItemAccessibleRole}` = "ListItem";
 
-	@property({ type: ListSelectionMode, defaultValue: ListSelectionMode.None })
-	_selectionMode!: `${ListSelectionMode}`;
+	@property()
+	_selectionMode: `${ListSelectionMode}` = "None";
 
 	/**
 	 * Defines the delete button, displayed in "Delete" mode.
@@ -232,13 +231,10 @@ abstract class ListItem extends ListItemBase {
 	}
 
 	onEnterDOM() {
+		super.onEnterDOM();
 		document.addEventListener("mouseup", this.deactivate);
 		document.addEventListener("touchend", this.deactivate);
 		document.addEventListener("keyup", this.deactivateByKey);
-
-		if (isDesktop()) {
-			this.setAttribute("desktop", "");
-		}
 	}
 
 	onExitDOM() {