+++ packages/main/src/ListItemCustom.ts
@@ -39,7 +39,7 @@ class ListItemCustom extends ListItem {
 	 * @since 2.0.0
 	 */
 	@property({ type: Boolean })
-	movable!: boolean;
+	movable = false;
 
 	/**
 	 * Defines the text alternative of the component.
@@ -50,7 +50,7 @@ class ListItemCustom extends ListItem {
 	 * @since 1.0.0-rc.15
 	 */
 	@property()
-	declare accessibleName: string;
+	declare accessibleName
 
 	async _onkeydown(e: KeyboardEvent) {
 		const isTab = isTabNext(e) || isTabPrevious(e);