+++ packages/main/src/SpecialCalendarDate.ts
@@ -1,7 +1,7 @@
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import CalendarDate from "./CalendarDate.js";
-import CalendarLegendItemType from "./types/CalendarLegendItemType.js";
+import type CalendarLegendItemType from "./types/CalendarLegendItemType.js";
 
 /**
  * @class
@@ -23,8 +23,8 @@ class SpecialCalendarDate extends CalendarDate {
 	 * @default "None"
 	 * @public
 	 */
-	@property({ type: CalendarLegendItemType, defaultValue: CalendarLegendItemType.None })
-	type!: `${CalendarLegendItemType}`;
+	@property()
+	type: `${CalendarLegendItemType}` = "None";
 }
 
 SpecialCalendarDate.define();