+++ packages/main/src/SuggestionItem.ts
@@ -2,9 +2,9 @@ import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 
-import ValueState from "@ui5/webcomponents-base/dist/types/ValueState.js";
+import type ValueState from "@ui5/webcomponents-base/dist/types/ValueState.js";
 import SuggestionListItem from "./SuggestionListItem.js";
-import ListItemType from "./types/ListItemType.js";
+import type ListItemType from "./types/ListItemType.js";
 import type { IInputSuggestionItem } from "./Input.js";
 
 /**
@@ -27,7 +27,7 @@ class SuggestionItem extends UI5Element implements IInputSuggestionItem {
 	 * @public
 	 */
 	@property()
-	text!: string
+	text = "";
 
 	/**
 	 * Defines the visual indication and behavior of the item.
@@ -39,8 +39,8 @@ class SuggestionItem extends UI5Element implements IInputSuggestionItem {
 	 * @public
 	 * @since 1.0.0-rc.8
 	*/
-	@property({ type: ListItemType, defaultValue: ListItemType.Active })
-	type!: `${ListItemType}`
+	@property()
+	type: `${ListItemType}` = "Active";
 
 	/**
 	 * Defines the description displayed right under the item text, if such is present.
@@ -48,7 +48,7 @@ class SuggestionItem extends UI5Element implements IInputSuggestionItem {
 	 * @public
 	 */
 	@property()
-	description!: string
+	description = "";
 
 	/**
 	 * Defines the `icon` source URI.
@@ -56,11 +56,11 @@ class SuggestionItem extends UI5Element implements IInputSuggestionItem {
 	 * **Note:**
 	 * SAP-icons font provides numerous built-in icons. To find all the available icons, see the
 	 * [Icon Explorer](https://sdk.openui5.org/test-resources/sap/m/demokit/iconExplorer/webapp/index.html).
-	 * @default ""
+	 * @default undefined
 	 * @public
 	 */
 	@property()
-	icon!: string
+	icon?: string;
 
 	/**
 	 * Defines whether the `icon` should be displayed in the beginning of the item or in the end.
@@ -70,16 +70,17 @@ class SuggestionItem extends UI5Element implements IInputSuggestionItem {
 	 * @public
 	 */
 	@property({ type: Boolean })
-	iconEnd!: boolean;
+	iconEnd = false;
+
 	/**
 	 * Defines the `image` source URI.
 	 *
 	 * **Note:** The `image` would be displayed in the beginning of the item.
-	 * @default ""
+	 * @default undefined
 	 * @public
 	 */
 	@property()
-	image!: string
+	image?: string;
 
 	/**
 	 * Defines the `additionalText`, displayed in the end of the item.
@@ -88,7 +89,7 @@ class SuggestionItem extends UI5Element implements IInputSuggestionItem {
 	 * @public
 	 */
 	@property()
-	additionalText!: string
+	additionalText = "";
 
 	/**
 	 * Defines the state of the `additionalText`.
@@ -96,8 +97,8 @@ class SuggestionItem extends UI5Element implements IInputSuggestionItem {
 	 * @since 1.0.0-rc.15
 	 * @public
 	 */
-	@property({ type: ValueState, defaultValue: ValueState.None })
-	additionalTextState!: `${ValueState}`
+	@property()
+	additionalTextState: `${ValueState}` = "None";
 
 	get groupItem() {
 		return false;