+++ packages/main/src/types/IconMode.ts
@@ -13,16 +13,16 @@ enum IconMode {
 
 	/**
 	 * Decorative mode.
-	 * Configures the component to internally render role="button".
-	 * This mode also supports focus and press handling to enhance interactivity.
+	 * Configures the component to internally render role="presentation" and aria-hidden="true",
+	 * making it purely decorative without semantic content or interactivity.
 	 * @public
 	 */
 	Decorative = "Decorative",
 
 	/**
 	 * Interactive mode.
-	 * Configures the component to internally render role="presentation" and aria-hidden="true",
-	 * making it purely decorative without semantic content or interactivity.
+	 * Configures the component to internally render role="button".
+	 * This mode also supports focus and press handling to enhance interactivity.
 	 * @public
 	*/
 	Interactive = "Interactive",